refactor: move 'decorators' folder to 'common'
