+++ backend/src/auth/auth.controller.ts
@@ -19,7 +19,7 @@ import { AuthGuard } from '@nestjs/passport';
 import { GetUsersFilterDto } from './dto/get-users-filter.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { setStorageOptions } from 'src/common/file-upload';
-import { GetUser } from 'src/decorators/get-user.decorator';
+import { GetUser } from 'src/common/decorators/get-user.decorator';
 import { UpdateCredentialsDto } from './dto/update-credentials.dto';
 const allowedExtensions: string[] = ['.jpg', '.jpeg', '.png'];
